mark library in physics plugin as depricated
